requirements.txt
astroid==2.9.3
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Faker==11.3.0
Flask==2.0.2
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
fonttools==4.29.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pycparser==2.21
pylint==2.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.3.24
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
